fix: Fix type mismatch in calculate_batch_progress method
- Cast ceil() results to integers to avoid float vs int comparison failures
- batches_completed now returns int 3 instead of float 3.0
- total_batches now returns int 4 instead of float 4.0
- Resolves test failure where batches_completed === 3 was failing due to type mismatch

+++ wp-content/plugins/vd-license-manager/includes/modules/utility-helper/components/class-calculation-helper.php
@@ -226,8 +226,8 @@ class CalculationHelper implements CalculationHelperInterface {
         }
 
         $percentage = self::calculate_percentage($processed, $total, 1);
-        $batches_completed = $batch_size > 0 ? ceil($processed / $batch_size) : 0;
-        $total_batches = $batch_size > 0 ? ceil($total / $batch_size) : 0;
+        $batches_completed = $batch_size > 0 ? (int) ceil($processed / $batch_size) : 0;
+        $total_batches = $batch_size > 0 ? (int) ceil($total / $batch_size) : 0;
         $batches_remaining = max(0, $total_batches - $batches_completed);
         $items_remaining = max(0, $total - $processed);
